In memory database research for testing purposes.

+++ back-end/tons-of-tacos/src/main/java/com/adamstraub/tonsoftacos/tonsoftacos/controller/BaseMenuItemController.java
@@ -3,14 +3,13 @@ package com.adamstraub.tonsoftacos.tonsoftacos.controller;
 import com.adamstraub.tonsoftacos.tonsoftacos.entities.MenuItem;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
 
 @RestController
 public class BaseMenuItemController implements MenuItemController {
 
 
     @Override
-    public MenuItem fetchMenuItem(Long id) {
+    public MenuItem fetchMenuItem(Integer id) {
         return null;
     }
 }
